210909 Macfa ] 방문 버그
비로그인 시에도
방문이력을 가져오는 문제 수정

+++ app/Models/Visit.php
@@ -22,23 +22,28 @@ class Visit extends Model
         return $this->belongsTo(User::class, 'userID', 'id');
     }
     public function addHistory(Channel $channel) {
-        $channel->visits()->updateOrCreate([
-            'userID' => auth()->id(),
-            'channelID' => $channel->id
-        ], [
-            'updated_at' => now()
-        ]);
+        if(auth()->check()) {
+            $channel->visits()->updateOrCreate([
+                'userID' => auth()->id(),
+                'channelID' => $channel->id
+            ], [
+                'updated_at' => now()
+            ]);
 
-        $totalCount = Visit::where('userID', auth()->id())->count();
-        // 최대 5개의 방문이력만 허용
-        if($totalCount >= 5) {
-            Visit::where('userID', auth()->id())->orderby('updated_at', 'asc')->limit(1)->delete();
+            $totalCount = Visit::where('userID', auth()->id())->count();
+            // 최대 5개의 방문이력만 허용
+            if($totalCount >= 5) {
+                Visit::where('userID', auth()->id())->orderby('updated_at', 'asc')->limit(1)->delete();
+            }
+
+            return $this->showHistory();
         }
-
-        return $this->showHistory();
     }
     public function showHistory() {
-        $user = User::find(auth()->id());
-        return $user->visits()->orderby('updated_at', 'desc')->get();
+        if(auth()->check() ) {
+            $user = User::find(auth()->id());
+
+            return $user->visits()->orderby('updated_at', 'desc')->get();
+        }
     }
 }
